+++ website/src/App.tsx
@@ -13,6 +13,7 @@ import { BestPractices } from './pages/docs/BestPractices'
 import { MCPIntegration } from './pages/docs/MCPIntegration'
 import { Terms } from './pages/Terms'
 import { Privacy } from './pages/Privacy'
+import { Changelog } from './pages/Changelog'
 
 function App() {
   return (
@@ -32,6 +33,7 @@ function App() {
             <Route path="/commands" element={<Commands />} />
             <Route path="/workflows" element={<Workflows />} />
             <Route path="/faq" element={<FAQPage />} />
+            <Route path="/changelog" element={<Changelog />} />
             <Route path="/terms" element={<Terms />} />
             <Route path="/privacy" element={<Privacy />} />
           </Routes>